requirements.txt
utils_gemini
IPython
csv
os
python-dotenv
google-generativeai